cpu/system: Improve auto-joypad poll timing, modify state handling
Backport: https://github.com/bsnes-emu/bsnes/pull/351 by Morilli

This commit breaks state compatibility with standalone bsnes. Previous
to this commit, the raw state data was still compatible, but would
require some manipulation on the user's part as the standalone creates
compressed states and encapsulates them in .bsz format. As standalone
bsnes also broke the state format for this change, it made sense to use
this opportunity to version the states differently in bsnes-jg. Effort
has been taken to allow loading states created in previous versions of
the emu, with the caveat that there may be some unexpected behaviour,
though this is unlikely in nearly all cases.